Add files via upload
This GitHub project implements a predictive model using Linear Regression to estimate insurance charges based on relevant features such as age, BMI, smoking habits, etc. Linear Regression is employed to create a model that can analyze the relationships between these factors and predict insurance charges accurately. The dataset used for training and testing the model is cleaned, preprocessed, and then utilized to train the Linear Regression model. Evaluation metrics are included to assess the model's performance and help users understand the Model accuracy and reliability. This project aims to provide a foundational understanding of predictive modeling in the insurance domain.